refactor some codes

+++ app/controllers/dashboard_controller.rb
@@ -5,8 +5,8 @@ class DashboardController < ApplicationController
   	@totalmembers = Membership.count
   	@memberfemale = Membership.where(gender: 'Female').count
   	@membermale = Membership.where(gender: 'Male').count
-  	@malepercent = @membermale.to_d / @totalmembers.to_d * 100
-  	@femalepercent = @memberfemale.to_d / @totalmembers.to_d * 100
+  	@malepercent = @membermale.to_f / @totalmembers.to_f * 100
+  	@femalepercent = @memberfemale.to_f / @totalmembers.to_f * 100
   end
 
 end
